chore: move Menu to UI and HamburgerMenu to common
