+++ src/components/Common/UI/Alert/index.ts
@@ -1,0 +1,1 @@
+export { Alert } from './Alert'
