+++ src/contexts/ComponentAdapter/adapters/internalComponentAdapter.tsx
@@ -1,5 +1,5 @@
-import { HamburgerMenu } from '@/components/Common/UI/HamburgerMenu'
-import type { HamburgerMenuProps } from '@/components/Common/UI/HamburgerMenu/HamburgerMenuTypes'
+import { HamburgerMenu } from '@/components/Common/HamburgerMenu'
+import type { HamburgerMenuProps } from '@/components/Common/HamburgerMenu/HamburgerMenuTypes'
 import type { InternalComponentsContextType } from '@/contexts/ComponentAdapter/useComponentContext'
 
 export const internalComponents: InternalComponentsContextType = {
